MM-29609 Move UnreadChannelIndicator into its own folder
